Simplify JniNativeLibrary code
The original implementation used three different
sets of P/Invoke declarations, so that the program
could detect and load the correct back-end C++ library
at runtime. I do not need this functionality; therefore,
I have trimmed the duplicate implementations.

Also renamed the DLL name that is being P/Invoked into
so that it reflects the name of the actual DLL on disk.